Add files via upload
Club_Supervisor
About
Club_Supervisor is a web application designed for efficient management and administration of clubs. It provides a comprehensive set of features to streamline club operations, facilitate communication, and enhance member engagement.

Technologies Used
Frontend Development:
HTML, CSS, JavaScript: These form the fundamental building blocks for creating the structure, style, and interactive elements of the web application.
Backend Development:
MySQL: Used for database management.
Features
Dashboard:
The Dashboard serves as the central hub of the Club Supervisor Project, providing an at-a-glance overview of critical information and key performance indicators.

Blog:
The Blog section of the Club Supervisor Project serves as a dynamic platform for sharing information, insights, and updates within the club community.

Club Notices:
The Club Notices section is a dedicated space for disseminating important announcements, updates, and notices.

Attendance:
The Attendance section of the Club Supervisor Project is a crucial tool for tracking and managing member participation in various club activities.

Member Management:
The Member Management section is the backbone of the project, allowing administrators to maintain a comprehensive database of club members.

Sessions:
The Sessions section is designed to organize and manage club sessions, workshops, or training programs.